+++ apps/platform.js
@@ -79,6 +79,7 @@ include(function () { //runs after prior includes are loaded
             "platform/user_context",
             "platform/tango_mixins",
             "platform/tango_webapp_storage"
+            // "platform/multimap"
         )
     }
 });
